% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFunctions.R
\name{convertHGNC}
\alias{convertHGNC}
\title{Rename features using HGNC symbols
Must be in following formats:
Transciptomics: XXXX/nENSGXXXXXXXXXXX
Proteomics: XXXXXX_XXXXXXXXXX_XXXXXX_(HGNC or NA)
Luminex Cytokine: Premade table}
\usage{
convertHGNC(model)
}
\arguments{
\item{model}{A DIABLO model object.}
}
\description{
Rename features using HGNC symbols
Must be in following formats:
Transciptomics: XXXX/nENSGXXXXXXXXXXX
Proteomics: XXXXXX_XXXXXXXXXX_XXXXXX_(HGNC or NA)
Luminex Cytokine: Premade table
}
