% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFunctions.R
\name{genesetEnrichment}
\alias{genesetEnrichment}
\title{Gene set enrichment analysis}
\usage{
genesetEnrichment(geneSet)
}
\arguments{
\item{geneSet}{A list of gene symbols.}
}
\description{
Gene set enrichment analysis
}
