% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ggProducePlot}
\alias{ggProducePlot}
\title{create the plot given a standard input of 2 data.frames
1. vars - a data.frame containing the variates
2. vecs (optional) - a data.frame of the loadings}
\usage{
ggProducePlot(vars, labs, vecs = NULL, topn = 10)
}
\description{
create the plot given a standard input of 2 data.frames
1. vars - a data.frame containing the variates
2. vecs (optional) - a data.frame of the loadings
}
