% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luminexcytokineMapping}
\alias{luminexcytokineMapping}
\title{Luminex Cytokine data mapping}
\format{TODO}
\source{
TODO
}
\usage{
data(luminexcytokineMapping)
}
\description{
Mappings from Luminex Cytokine data to HGNC symbols
}
\examples{
data(luminexcytokineMapping)
}
\references{
TODO
}
\keyword{datasets}
