% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFunctions.R
\name{matchPPI}
\alias{matchPPI}
\title{Match with existing Protein-Protein Interaction data
If there is a match, returns the index of the match in the PPI data in a vector of length 'edges'
'edges' and 'data' both two column data frames}
\usage{
matchPPI(edges, data)
}
\arguments{
\item{edges}{A data frame consisting of two proteins that have an edge.}

\item{data}{A data frame consisting of two columns of proteins that interact}
}
\description{
Match with existing Protein-Protein Interaction data
If there is a match, returns the index of the match in the PPI data in a vector of length 'edges'
'edges' and 'data' both two column data frames
}
