% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPIList}
\alias{PPIList}
\title{List of Protein-Protein Interactions}
\format{TODO}
\source{
TODO
}
\usage{
data(PPIList)
}
\description{
List of Protein-Protein Interactions
}
\examples{
data(PPIList)
}
\references{
TODO
}
\keyword{datasets}
