% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteinMapping}
\alias{proteinMapping}
\title{Protein data mapping}
\format{TODO}
\source{
TODO
}
\usage{
data(proteinMapping)
}
\description{
Mappings from Protein data to HGNC symbols
}
\examples{
data(proteinMapping)
}
\references{
TODO
}
\keyword{datasets}
