% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFunctions.R
\name{genesetEnrichment}
\alias{genesetEnrichment}
\title{Gene set enrichment analysis}
\usage{
genesetEnrichment(geneSet, featureMapping)
}
\arguments{
\item{geneSet}{A list of gene symbols.}

\item{featureMapping}{A list of data frames containing 'Data.Names', 'Gene.Symbols', 'Display.Names' for each of the data blocks.}
}
\description{
Gene set enrichment analysis
}
