% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize DIABLO models in an interactive environment}
\usage{
visualize(model, featureMapping = NULL)
}
\arguments{
\item{model}{A DIABLO model object.}

\item{featureMapping}{A list of data frames containing 'Data.Names', 'Gene.Symbols', 'Display.Names' for each of the data blocks.}
}
\description{
Visualize DIABLO models in an interactive environment
}
\author{
Jamie C. Ye <jamiec.ye@gmail.com>
}
